backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.31.0
motor>=3.6.0
pydantic>=2.9.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
anthropic>=0.45.0
python-dotenv>=1.0.0
email-validator>=2.2.0
python-multipart>=0.0.20
bleach>=6.0.0
fpdf2>=2.7.9

httpx
apscheduler>=3.10.0
pytz
python-dateutil

# ML dependencies
scikit-learn
joblib
numpy
pandas
